Removing extensions that exists on ktx
