+++ app/src/main/java/dk/nodes/template/util/Extensions.kt
@@ -1,0 +1,5 @@
+package dk.nodes.template.util
+
+inline fun <T> T.guard(block: T.() -> Unit): T {
+    if (this == null) block(); return this
+}
