+++ app/src/main/java/dk/nodes/template/util/LiveDataExtensions.kt
@@ -1,0 +1,32 @@
+package dk.nodes.template.util
+
+import androidx.lifecycle.LifecycleOwner
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.Observer
+
+inline fun <T> LiveData<T>.observe(
+    lifecycleOwner: LifecycleOwner,
+    crossinline observer: (T?) -> Unit
+) {
+    this.observe(lifecycleOwner, Observer {
+        observer(it)
+    })
+}
+
+inline fun <T> LiveData<T>.observeNonNull(
+    lifecycleOwner: LifecycleOwner,
+    crossinline observer: (T) -> Unit
+) {
+    this.observe(lifecycleOwner, Observer {
+        it?.let(observer)
+    })
+}
+
+inline fun <E, T : Event<E>> LiveData<T>.observeEvent(
+    lifecycleOwner: LifecycleOwner,
+    crossinline observer: (E) -> Unit
+) {
+    this.observe(lifecycleOwner, EventObserver {
+        observer(it)
+    })
+}
